backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
pydantic
supabase
python-jose[cryptography]
passlib[bcrypt]
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
beautifulsoup4
playwright
lxml
schedule
pydantic[email]
pydantic-settings
python-multipart
httpx
sqlalchemy-libsql
email-validator
chromadb
sentence-transformers
numpy
tqdm
